+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ## [unreleased]
 
+### Fixed
+- Missing `Custom style` button in Print label task [#2764]
+
+[#2764]: https://github.com/SpeciesFileGroup/taxonworks/issues/2764
 \-
 
 ## [0.22.6] - 2022-01-10